Moved dialogs to GeneralDialogCreation
