+++ src/main/java/com/amaze/filemanager/ui/dialogs/SmbConnectDialog.java
@@ -25,7 +25,6 @@ import com.amaze.filemanager.R;
 import com.amaze.filemanager.filesystem.HFile;
 import com.amaze.filemanager.utils.CryptUtil;
 import com.amaze.filemanager.utils.EditTextColorStateUtil;
-import com.amaze.filemanager.utils.Futils;
 import com.amaze.filemanager.utils.OpenMode;
 import com.amaze.filemanager.utils.PreferenceUtils;
 import com.amaze.filemanager.utils.Utils;
@@ -203,7 +202,7 @@ public class SmbConnectDialog extends DialogFragment {
         help.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Futils.showSMBHelpDialog(context,PreferenceUtils.getAccentString(sharedPreferences));
+                GeneralDialogCreation.showSMBHelpDialog(context,PreferenceUtils.getAccentString(sharedPreferences));
             }
         });
